+++ dvr_scan/shared/__init__.py
@@ -208,7 +208,6 @@ def init_scanner(
         min_event_len=settings.get("min-event-length"),
         time_pre_event=settings.get("time-before-event"),
         time_post_event=settings.get("time-post-event"),
-        use_pts=settings.get("use-pts"),
     )
 
     scanner.set_thumbnail_params(